Dissolve metadata resolver util into the idiomatic selector reads
